+++ vcr/credential/test.go
@@ -97,28 +97,6 @@ func JWTNutsOrganizationCredential(t *testing.T, subjectID did.DID) vc.Verifiabl
 	return *jwtVC
 }
 
-func ValidStatusList2021Credential(_ testing.TB) vc.VerifiableCredential {
-	id := ssi.MustParseURI("https://example.com/credentials/status/3")
-	validFrom := time.Now()
-	validUntilTomorrow := validFrom.Add(24 * time.Hour)
-	return vc.VerifiableCredential{
-		Context:          []ssi.URI{vc.VCContextV1URI(), statusList2021ContextURI},
-		ID:               &id,
-		Type:             []ssi.URI{vc.VerifiableCredentialTypeV1URI(), stringToURI(StatusList2021CredentialType)},
-		Issuer:           ssi.MustParseURI("did:example:12345"),
-		ValidFrom:        &validFrom,
-		ValidUntil:       &validUntilTomorrow,
-		CredentialStatus: nil,
-		CredentialSubject: []any{&StatusList2021CredentialSubject{
-			Id:            "https://example-com/status/3#list",
-			Type:          StatusList2021CredentialSubjectType,
-			StatusPurpose: "revocation",
-			EncodedList:   "H4sIAAAAAAAA_-zAsQAAAAACsNDypwqjZ2sAAAAAAAAAAAAAAAAAAACAtwUAAP__NxdfzQBAAAA=", // has bit 1 set to true
-		}},
-		Proof: []any{vc.Proof{}},
-	}
-}
-
 func stringToURI(input string) ssi.URI {
 	return ssi.MustParseURI(input)
 }
