+++ vcr/credential/validator.go
@@ -24,6 +24,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/nuts-foundation/nuts-node/vcr/statuslist"
 	"net/url"
 	"strconv"
 	"strings"
@@ -146,16 +147,16 @@ func validateCredentialStatus(credential vc.VerifiableCredential) error {
 		}
 
 		// only accept StatusList2021EntryType for now
-		if credentialStatus.Type != StatusList2021EntryType {
+		if credentialStatus.Type != statuslist.StatusList2021EntryType {
 			continue
 		}
 
-		if !credential.ContainsContext(statusList2021ContextURI) {
+		if !credential.ContainsContext(statuslist.statusList2021ContextURI) {
 			return errors.New("StatusList2021 context is required")
 		}
 
 		// unmarshal as StatusList2021Entry
-		var cs StatusList2021Entry
+		var cs statuslist.StatusList2021Entry
 		if err = json.Unmarshal(credentialStatus.Raw(), &cs); err != nil {
 			return err
 		}
@@ -318,49 +319,5 @@ func validateNutsCredentialID(credential vc.VerifiableCredential) error {
 	if id.String() != credential.Issuer.String() {
 		return failure("credential ID must start with issuer")
 	}
-	return nil
-}
-
-// statusList2021CredentialValidator validates that all required fields of a StatusList2021CredentialType are present
-type statusList2021CredentialValidator struct{}
-
-func (d statusList2021CredentialValidator) Validate(credential vc.VerifiableCredential) error {
-	if err := (defaultCredentialValidator{}).Validate(credential); err != nil {
-		return err
-	}
-
-	{ // Credential checks
-		if !credential.ContainsContext(statusList2021ContextURI) {
-			return failure("context '%s' is required", statusList2021ContextURI)
-		}
-		if !credential.IsType(statusList2021CredentialTypeURI) {
-			return failure("type '%s' is required", statusList2021CredentialTypeURI)
-		}
-	}
-
-	{ // CredentialSubject checks
-		var target []StatusList2021CredentialSubject
-		err := credential.UnmarshalCredentialSubject(&target)
-		if err != nil {
-			return failure(err.Error())
-		}
-		// The spec is not clear if there could be multiple CredentialSubjects. This could allow 'revocation' and 'suspension' to be defined in a single credential.
-		// However, it is not defined how to select the correct list (StatusPurpose) when validating credentials that are using this StatusList2021Credential.
-		if len(target) != 1 {
-			return failure("single CredentialSubject expected")
-		}
-		cs := target[0]
-
-		if cs.Type != StatusList2021CredentialSubjectType {
-			return failure("credentialSubject.type '%s' is required", StatusList2021CredentialSubjectType)
-		}
-		if cs.StatusPurpose == "" {
-			return failure("credentialSubject.statusPurpose is required")
-		}
-		if cs.EncodedList == "" {
-			return failure("credentialSubject.encodedList is required")
-		}
-	}
-
 	return nil
 }
