+++ vcr/credential/validator_test.go
@@ -23,6 +23,7 @@ import (
 	ssi "github.com/nuts-foundation/go-did"
 	"github.com/nuts-foundation/go-did/vc"
 	"github.com/nuts-foundation/nuts-node/jsonld"
+	"github.com/nuts-foundation/nuts-node/vcr/statuslist"
 	"github.com/nuts-foundation/nuts-node/vdr"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
@@ -509,68 +510,6 @@ func TestDefaultCredentialValidator(t *testing.T) {
 	})
 }
 
-func TestStatusList2021CredentialValidator_Validate(t *testing.T) {
-	t.Run("ok", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.NoError(t, err)
-	})
-	t.Run("error - wraps defaultCredentialValidator", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.Context = []ssi.URI{statusList2021CredentialTypeURI}
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: default context is required")
-	})
-	t.Run("error - missing status list context", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.Context = []ssi.URI{vc.VCContextV1URI()}
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: context 'https://w3id.org/vc/status-list/2021/v1' is required")
-	})
-	t.Run("error - missing StatusList credential type", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.Type = []ssi.URI{vc.VerifiableCredentialTypeV1URI()}
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: type 'StatusList2021Credential' is required")
-	})
-	t.Run("error - invalid credential subject", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.CredentialSubject = []any{"{"}
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: json: cannot unmarshal string into Go value of type credential.StatusList2021CredentialSubject")
-	})
-	t.Run("error - wrong credential subject", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.CredentialSubject = []any{NutsAuthorizationCredentialSubject{}}
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: credentialSubject.type 'StatusList2021' is required")
-	})
-	t.Run("error - multiple credentialSubject", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.CredentialSubject = []any{StatusList2021CredentialSubject{}, StatusList2021CredentialSubject{}}
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: single CredentialSubject expected")
-	})
-	t.Run("error - missing credentialSubject.type", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.CredentialSubject[0].(*StatusList2021CredentialSubject).Type = ""
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: credentialSubject.type 'StatusList2021' is required")
-	})
-	t.Run("error - missing statusPurpose", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.CredentialSubject[0].(*StatusList2021CredentialSubject).StatusPurpose = ""
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: credentialSubject.statusPurpose is required")
-	})
-	t.Run("error - missing encodedList", func(t *testing.T) {
-		cred := ValidStatusList2021Credential(t)
-		cred.CredentialSubject[0].(*StatusList2021CredentialSubject).EncodedList = ""
-		err := statusList2021CredentialValidator{}.Validate(cred)
-		assert.EqualError(t, err, "validation failed: credentialSubject.encodedList is required")
-	})
-}
-
 func Test_validateCredentialStatus(t *testing.T) {
 	t.Run("ok - no credentialStatus", func(t *testing.T) {
 		assert.NoError(t, validateCredentialStatus(vc.VerifiableCredential{}))
@@ -591,13 +530,13 @@ func Test_validateCredentialStatus(t *testing.T) {
 		assert.EqualError(t, err, "credentialStatus.type is required")
 	})
 
-	t.Run(StatusList2021EntryType, func(t *testing.T) {
+	t.Run(statuslist.StatusList2021EntryType, func(t *testing.T) {
 		makeValidCSEntry := func() vc.VerifiableCredential {
 			return vc.VerifiableCredential{
-				Context: []ssi.URI{statusList2021ContextURI},
-				CredentialStatus: []any{&StatusList2021Entry{
+				Context: []ssi.URI{statuslist.statusList2021ContextURI},
+				CredentialStatus: []any{&statuslist.StatusList2021Entry{
 					ID:                   "https://example-com/credentials/status/3#94567",
-					Type:                 StatusList2021EntryType,
+					Type:                 statuslist.StatusList2021EntryType,
 					StatusPurpose:        "revocation",
 					StatusListIndex:      "94567",
 					StatusListCredential: "https://example-com/credentials/status/3",
@@ -621,31 +560,31 @@ func Test_validateCredentialStatus(t *testing.T) {
 		})
 		t.Run("error - id == statusListCredential", func(t *testing.T) {
 			cred := makeValidCSEntry()
-			cred.CredentialStatus[0].(*StatusList2021Entry).ID = cred.CredentialStatus[0].(*StatusList2021Entry).StatusListCredential
+			cred.CredentialStatus[0].(*statuslist.StatusList2021Entry).ID = cred.CredentialStatus[0].(*statuslist.StatusList2021Entry).StatusListCredential
 			err := validateCredentialStatus(cred)
 			assert.EqualError(t, err, "StatusList2021Entry.id is the same as the StatusList2021Entry.statusListCredential")
 		})
 		t.Run("error - missing statusPurpose", func(t *testing.T) {
 			cred := makeValidCSEntry()
-			cred.CredentialStatus[0].(*StatusList2021Entry).StatusPurpose = ""
+			cred.CredentialStatus[0].(*statuslist.StatusList2021Entry).StatusPurpose = ""
 			err := validateCredentialStatus(cred)
 			assert.EqualError(t, err, "StatusList2021Entry.statusPurpose is required")
 		})
 		t.Run("error - statusListIndex is negative", func(t *testing.T) {
 			cred := makeValidCSEntry()
-			cred.CredentialStatus[0].(*StatusList2021Entry).StatusListIndex = "-1"
+			cred.CredentialStatus[0].(*statuslist.StatusList2021Entry).StatusListIndex = "-1"
 			err := validateCredentialStatus(cred)
 			assert.EqualError(t, err, "invalid StatusList2021Entry.statusListIndex")
 		})
 		t.Run("error - statusListIndex is not a number", func(t *testing.T) {
 			cred := makeValidCSEntry()
-			cred.CredentialStatus[0].(*StatusList2021Entry).StatusListIndex = "one"
+			cred.CredentialStatus[0].(*statuslist.StatusList2021Entry).StatusListIndex = "one"
 			err := validateCredentialStatus(cred)
 			assert.EqualError(t, err, "invalid StatusList2021Entry.statusListIndex")
 		})
 		t.Run("error - statusListCredential is not a valid URL", func(t *testing.T) {
 			cred := makeValidCSEntry()
-			cred.CredentialStatus[0].(*StatusList2021Entry).StatusListCredential = "not a URL"
+			cred.CredentialStatus[0].(*statuslist.StatusList2021Entry).StatusListCredential = "not a URL"
 			err := validateCredentialStatus(cred)
 			assert.EqualError(t, err, "parse StatusList2021Entry.statusListCredential URL: parse \"not a URL\": invalid URI for request")
 		})
