+++ vcr/statuslist/credential_status_test.go
@@ -16,11 +16,12 @@
  *
  */
 
-package verifier
+package statuslist
 
 import (
 	"encoding/json"
 	"errors"
+	"github.com/nuts-foundation/nuts-node/vcr/verifier"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -84,8 +85,8 @@ func TestCredentialStatus_verify(t *testing.T) {
 	})
 	t.Run("error - statusPurpose mismatch", func(t *testing.T) {
 		// server that return StatusList2021Credential with statusPurpose == suspension
-		statusList2021Credential := credential.ValidStatusList2021Credential(t)
-		statusList2021Credential.CredentialSubject[0].(*credential.StatusList2021CredentialSubject).StatusPurpose = "suspension"
+		statusList2021Credential := ValidStatusList2021Credential(t)
+		statusList2021Credential.CredentialSubject[0].(*StatusList2021CredentialSubject).StatusPurpose = "suspension"
 		credBytes, err := json.Marshal(statusList2021Credential)
 		require.NoError(t, err)
 		ts := httptest.NewTLSServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
@@ -143,7 +144,7 @@ func TestCredentialStatus_update(t *testing.T) {
 	})
 	t.Run("error - verifyStatusList2021Credential", func(t *testing.T) {
 		cs, _, ts := testSetup(t, false)
-		mockVerifier := NewMockVerifier(gomock.NewController(t))
+		mockVerifier := verifier.NewMockVerifier(gomock.NewController(t))
 		mockVerifier.EXPECT().VerifySignature(gomock.Any(), nil).Return(errors.New("custom error"))
 		cs.verifySignature = mockVerifier.VerifySignature
 
@@ -156,7 +157,7 @@ func TestCredentialStatus_update(t *testing.T) {
 
 func TestCredentialStatus_download(t *testing.T) {
 	t.Run("ok", func(t *testing.T) {
-		cred := credential.ValidStatusList2021Credential(t) // has bit 1 set
+		cred := ValidStatusList2021Credential(t) // has bit 1 set
 		expected, err := json.Marshal(cred)
 		require.NoError(t, err)
 		ts := httptest.NewTLSServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
@@ -216,8 +217,8 @@ func TestCredentialStatus_verifyStatusList2021Credential(t *testing.T) {
 		},
 	}
 	t.Run("ok", func(t *testing.T) {
-		cred := credential.ValidStatusList2021Credential(t)
-		expected := cred.CredentialSubject[0].(*credential.StatusList2021CredentialSubject)
+		cred := ValidStatusList2021Credential(t)
+		expected := cred.CredentialSubject[0].(*StatusList2021CredentialSubject)
 		credSubj, err := credentialStatusNoSignCheck.verifyStatusList2021Credential(cred)
 		assert.NoError(t, err)
 		require.NotNil(t, credSubj)
@@ -230,37 +231,37 @@ func TestCredentialStatus_verifyStatusList2021Credential(t *testing.T) {
 		assert.Nil(t, credSubj)
 	})
 	t.Run("error - too many credential types", func(t *testing.T) {
-		cred := credential.ValidStatusList2021Credential(t)
+		cred := ValidStatusList2021Credential(t)
 		cred.Type = append(cred.Type, ssi.MustParseURI("OneTooMany"))
 		credSubj, err := credentialStatusNoSignCheck.verifyStatusList2021Credential(cred)
 		assert.EqualError(t, err, "incorrect credential types")
 		assert.Nil(t, credSubj)
 	})
 	t.Run("error - credential validation failed", func(t *testing.T) {
-		cred := credential.ValidStatusList2021Credential(t)
-		cred.CredentialSubject[0].(*credential.StatusList2021CredentialSubject).Type = "wrong type"
+		cred := ValidStatusList2021Credential(t)
+		cred.CredentialSubject[0].(*StatusList2021CredentialSubject).Type = "wrong type"
 		credSubj, err := credentialStatusNoSignCheck.verifyStatusList2021Credential(cred)
 		assert.EqualError(t, err, "validation failed: credentialSubject.type 'StatusList2021' is required")
 		assert.Nil(t, credSubj)
 	})
 	t.Run("error - contains CredentialStatus", func(t *testing.T) {
-		cred := credential.ValidStatusList2021Credential(t)
+		cred := ValidStatusList2021Credential(t)
 		cred.CredentialStatus = []any{}
 		credSubj, err := credentialStatusNoSignCheck.verifyStatusList2021Credential(cred)
 		assert.EqualError(t, err, "StatusList2021Credential with a CredentialStatus is not supported")
 		assert.Nil(t, credSubj)
 	})
 	t.Run("error - invalid credentialSubject.encodedList", func(t *testing.T) {
-		cred := credential.ValidStatusList2021Credential(t)
-		cred.CredentialSubject[0].(*credential.StatusList2021CredentialSubject).EncodedList = "@"
+		cred := ValidStatusList2021Credential(t)
+		cred.CredentialSubject[0].(*StatusList2021CredentialSubject).EncodedList = "@"
 		credSubj, err := credentialStatusNoSignCheck.verifyStatusList2021Credential(cred)
 
 		assert.EqualError(t, err, "credentialSubject.encodedList is invalid: illegal base64 data at input byte 0")
 		assert.Nil(t, credSubj)
 	})
 	t.Run("error -invalid signature", func(t *testing.T) {
-		cred := credential.ValidStatusList2021Credential(t)
-		mockVerifier := NewMockVerifier(gomock.NewController(t))
+		cred := ValidStatusList2021Credential(t)
+		mockVerifier := verifier.NewMockVerifier(gomock.NewController(t))
 		mockVerifier.EXPECT().VerifySignature(cred, nil).Return(errors.New("invalid signature"))
 		cs := credentialStatus{verifySignature: mockVerifier.VerifySignature}
 		credSubj, err := cs.verifyStatusList2021Credential(cred)
@@ -273,9 +274,9 @@ func TestCredentialStatus_verifyStatusList2021Credential(t *testing.T) {
 //   - credentialStatus that does NOT verify signatures, and a client configured for the test server
 //   - a StatusList2021Entry pointing to the test server, optionally provide a statusListIndex matching statusList2021Credential.encodedList to simulate revocation
 //   - the test server
-func testSetup(t testing.TB, entryIsRevoked bool) (*credentialStatus, credential.StatusList2021Entry, *httptest.Server) {
+func testSetup(t testing.TB, entryIsRevoked bool) (*credentialStatus, StatusList2021Entry, *httptest.Server) {
 	// make test server
-	statusList2021Credential := credential.ValidStatusList2021Credential(t) // has bit 1 set
+	statusList2021Credential := ValidStatusList2021Credential(t) // has bit 1 set
 	credBytes, err := json.Marshal(statusList2021Credential)
 	if err != nil {
 		t.Fatal(err)
@@ -296,8 +297,8 @@ func testSetup(t testing.TB, entryIsRevoked bool) (*credentialStatus, credential
 	}
 
 	// make StatusList2021Entry
-	slEntry := credential.StatusList2021Entry{
-		Type:                 credential.StatusList2021EntryType,
+	slEntry := StatusList2021Entry{
+		Type:                 StatusList2021EntryType,
 		StatusPurpose:        "revocation",
 		StatusListIndex:      "76248",
 		StatusListCredential: ts.URL,
